Heroku: Fix Github Workflow Action extract_client
Remove dangling commits unrelated to client/.

Sometimes
    git subtree split --prefix=client -b client
is based on a dangling commit that is no longer referenced on any
branch which prevents this command from doing the expected job,
it does quite the opposite in fact and we end up with the reular tree
(back + front/client).

The workaround I found is to use
    git filter-branch --subdirectory-filter client
which removes the commits unrelated to the client/ folder.

Sources:
- https://medium.com/@porteneuve/mastering-git-subtrees-943d29a798ec#62d3
- https://stackoverflow.com/questions/5760331/git-subtree-is-not-retaining-history-so-i-cannot-push-subtree-changes-how-can-i